parking_backend/requirements.txt
ansible==2.10.3
ansible-base==2.10.3
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
factory-boy==3.1.0
Faker==4.17.0
freezegun==1.0.0
idna==2.10
install==1.3.4
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.7
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
